Version 1.1 - Phys-Spec Split
Completion of the physical-special split has been attained.

Move names in the summary screen now have a color that corresponds to
the move category they belong to (physical, special, status) for the
physical-special split feature.

Updated road map to reflect completion, in addition to adding/changing a
few minor things.