Version 1.4 - Chain Fishing
Consecutive successful reels leads to a higher chance of a shiny
encounter. This feature can be enabled using the `CHAIN_FISHING` define
in the config header.

Also added some code for the Shiny Charm, which will be implemented at a
later time. This was done mainly because the chain fishing tutorial
includes it.