Reduction in Flag Quantity
Sharply reduced the number of flags used by the game, and made it where
they can dynamically reassign themselves to a new address. This was done
using A LOT of define tags. This change can only be active if both
`PROJECT_VERDANT` and `EMER_REDUCED` are enabled in the config file.

These changes are not vanilla compatible (well, techinically they are,
but it'll REALLY mess up the events in your save data).