Poryscript Support, New Patch Ver., Updated Roadmap
A lot of little changes in this commit that mean a lot for the project.

First off, added support for Poryscript. Using this as opposed to the
standard .inc files is a lot more developer friendly and will help with
progressing the project at a (hopefully) quicker rate.

Next, bumped the current version to 2.4.1 over 2.4.1. While
_technically_ no patch was actually implemented, with all of the
integrations from upstream, I figure it would be best to at least change
the version to reflect this.

Finally, the roadmap has been adjusted. Each version has a to-do list of
items as opposed to a specific subversion tied to the goal. In addition,
there has been a slight change in the version categories and new content
has branched off into feature porting and new content.

Signed-off-by: wheeler-cs <wheelercs@protonmail.com>