Fixed Held Item Use Glitch
There was an issue with a malformed animation playing when a held item
was used in battle. Didn't trace down the issue to _exactly_ what it
was, but it was something to do with an attempted optimization with the
`AffineAnimCmd` union. Things have been fixed with regards to it, and
everything seems to be in working order... for now.

Note to self: Just because something is unused in code doesn't
necessarily mean it can be removed w/o worry.

Signed-off-by: wheeler-cs <wheelercs@protonmail.com>