Patch 3.0.2 - Forgotten Increment and Shiny Values
Forgot to increment to 3.0.1 last commit, which was a quick 1-line
change.

Added an `INCREASED_SHINY_ODDS` macro that can be used to set a custom
shiny odds. By default, it is set to gen VI's shiny rate (1/4096), but
this is easily modified by changing the value in `pokemon.h`.

Signed-off-by: wheeler-cs <wheelercs@protonmail.com>