Fixed Palette Bug with Battle Menu
The new balls had particle tags that collided with tags intended to be
part of the UI in a battle. As such, the elements in a battle were
glitched. This should be fixed, but it needs to be double-checked that
the new values do not collide with any other tags.

Signed-off-by: wheeler-cs <wheelercs@protonmail.com>