Cherish Ball Implementation
The cherish ball is [functionally] implemented and is about 95% working.
There is an issue when viewing a Pokemon in the summary screen where if
it was caught using a cherish ball, the regular Pokeball is displayed
instead. Don't know why, as the animation in battle currently that uses
the same sprite works. It's just in the summary screen it doesn't work.

Signed-off-by: wheeler-cs <wheelercs@protonmail.com>