Fixed New Balls Not Working
Issue was related to the numeric value of a Pokeball for a Pokemon being
stored as 4 bits in the structure. Did some fairly substantial
refactoring to the game's codebase to fix this, including:

- Removal of all defines for gen 4 games and reduction of the
  VERSION_GAMECUDE value down to 6 to fit the "game met" field into 3
  bits instead of 4.

- Movement of the fields within the PokemonSubstruct3 to make sure
  everything was word-aligned while also ensuring it fit in the original
  12 bytes all the other structures fit in.

- Allocated an extra bit to the "ability number" field to allow for
  implementation of hidden abilities at a future date.

**This is a game-breaking change.** Pokemon data for saves from previous
versions will no longer work properly. Most of the data is preservered,
but a lot is not.

Signed-off-by: wheeler-cs <wheelercs@protonmail.com>