Giving Go-Ahead w/ Ball Completion
Multipliers for newly-added Pokeballs appear to be working properly.
Officially giving approval for it being a completed feature.

Added `PV_DEBUG` compile option, which is useful in debugging various features
within the code base during runtime.

Updated the roadmap, which had a lot of errors and was very out of date.

Signed-off-by: wheeler-cs <wheelercs@protonmail.com>