Work on Petaburg Woods Extension
More work on the exterior of the new Petalburg Woods area. This includes
map design and encounter tables.

Also working on the interior of the house on this map.

Create an `EWRAM` variable that will be used for the interior of the
house for this map as well. The new variable `gForceEnableEncounters`
will be used to force encounters to occur, even on non-encounter tiles.

Signed-off-by: wheeler-cs <wheelercs@protonmail.com>