Working on Haunted Mansion
Working on the haunted mansion in Petalburg Woods. Spent a lot of time
getting the hidden boss event to work the way I wanted it too, and,
along the way, have had some other things to work on. This includes
several bug fixes.

The mansion is not complete yet, but the (hopefully) more difficult
stuff is behind me.

Signed-off-by: wheeler-cs <wheelercs@protonmail.com>